% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybiplot.R
\name{mybiplot}
\alias{mybiplot}
\title{Analyse attribute for outliers with respect to k}
\usage{
mybiplot(x = double(), k = double())
}
\arguments{
\item{x}{A numeric vector of continuous values.}

\item{k}{A numeric constant indicating the number of standard deviations to define outliers.}
}
\value{
A list containing:
\describe{
\item{result_table}{A data frame with original values, z-scores, and in/out classification.}
\item{plot}{A histogram ggplot object.}
\item{summary}{A summary table with statistics grouped by in/out classification.}
}
}
\description{
Analyse attribute for outliers with respect to k
}
\examples{
mybiplot(rnorm(100), 2)
}
